+++ rust/peerbeat_core/src/frb_generated.rs
@@ -38,7 +38,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.12.0";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -611884769;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = 1328805397;
 
 // Section: executor
 
@@ -1565,6 +1565,43 @@ fn wire__crate__api__library__library_set_favorite_impl(
                 transform_result_sse::<_, String>((move || {
                     let output_ok =
                         crate::api::library::library_set_favorite(api_track_id, api_on)?;
+                    Ok(output_ok)
+                })())
+            }
+        },
+    )
+}
+fn wire__crate__api__library__library_set_folder_watched_impl(
+    port_: flutter_rust_bridge::for_generated::MessagePort,
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_normal::<flutter_rust_bridge::for_generated::SseCodec, _, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "library_set_folder_watched",
+            port: Some(port_),
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Normal,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_folder_id = <i64>::sse_decode(&mut deserializer);
+            let api_watched = <bool>::sse_decode(&mut deserializer);
+            deserializer.end();
+            move |context| {
+                transform_result_sse::<_, String>((move || {
+                    let output_ok = crate::api::library::library_set_folder_watched(
+                        api_folder_id,
+                        api_watched,
+                    )?;
                     Ok(output_ok)
                 })())
             }
@@ -4160,118 +4197,124 @@ fn pde_ffi_dispatcher_primary_impl(
         47 => {
             wire__crate__api__library__library_set_favorite_impl(port, ptr, rust_vec_len, data_len)
         }
-        48 => wire__crate__api__library__library_start_watching_impl(
+        48 => wire__crate__api__library__library_set_folder_watched_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        49 => {
+        49 => wire__crate__api__library__library_start_watching_impl(
+            port,
+            ptr,
+            rust_vec_len,
+            data_len,
+        ),
+        50 => {
             wire__crate__api__library__library_stop_watching_impl(port, ptr, rust_vec_len, data_len)
         }
-        50 => {
+        51 => {
             wire__crate__api__library__library_track_by_id_impl(port, ptr, rust_vec_len, data_len)
         }
-        51 => {
+        52 => {
             wire__crate__api__library__library_track_count_impl(port, ptr, rust_vec_len, data_len)
         }
-        52 => {
+        53 => {
             wire__crate__api__library__library_track_lyrics_impl(port, ptr, rust_vec_len, data_len)
         }
-        53 => wire__crate__api__library__library_track_tags_impl(port, ptr, rust_vec_len, data_len),
-        54 => wire__crate__api__library__library_tracks_by_year_impl(
+        54 => wire__crate__api__library__library_track_tags_impl(port, ptr, rust_vec_len, data_len),
+        55 => wire__crate__api__library__library_tracks_by_year_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        55 => {
+        56 => {
             wire__crate__api__library__library_update_tags_impl(port, ptr, rust_vec_len, data_len)
         }
-        56 => wire__crate__api__network__net_active_peers_impl(port, ptr, rust_vec_len, data_len),
-        57 => wire__crate__api__library__net_clear_activity_impl(port, ptr, rust_vec_len, data_len),
-        58 => wire__crate__api__network__net_decide_peer_impl(port, ptr, rust_vec_len, data_len),
-        59 => wire__crate__api__network__net_discover_impl(port, ptr, rust_vec_len, data_len),
-        60 => wire__crate__api__library__net_forget_host_impl(port, ptr, rust_vec_len, data_len),
-        61 => wire__crate__api__network__net_host_port_impl(port, ptr, rust_vec_len, data_len),
-        62 => wire__crate__api__network__net_is_hosting_impl(port, ptr, rust_vec_len, data_len),
-        63 => wire__crate__api__library__net_known_fingerprints_impl(
+        57 => wire__crate__api__network__net_active_peers_impl(port, ptr, rust_vec_len, data_len),
+        58 => wire__crate__api__library__net_clear_activity_impl(port, ptr, rust_vec_len, data_len),
+        59 => wire__crate__api__network__net_decide_peer_impl(port, ptr, rust_vec_len, data_len),
+        60 => wire__crate__api__network__net_discover_impl(port, ptr, rust_vec_len, data_len),
+        61 => wire__crate__api__library__net_forget_host_impl(port, ptr, rust_vec_len, data_len),
+        62 => wire__crate__api__network__net_host_port_impl(port, ptr, rust_vec_len, data_len),
+        63 => wire__crate__api__network__net_is_hosting_impl(port, ptr, rust_vec_len, data_len),
+        64 => wire__crate__api__library__net_known_fingerprints_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        64 => wire__crate__api__library__net_known_host_fingerprint_impl(
+        65 => wire__crate__api__library__net_known_host_fingerprint_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        65 => wire__crate__api__network__net_party_active_impl(port, ptr, rust_vec_len, data_len),
-        66 => wire__crate__api__network__net_party_requests_impl(port, ptr, rust_vec_len, data_len),
-        67 => wire__crate__api__network__net_party_start_impl(port, ptr, rust_vec_len, data_len),
-        68 => wire__crate__api__network__net_party_stop_impl(port, ptr, rust_vec_len, data_len),
-        69 => wire__crate__api__network__net_party_update_impl(port, ptr, rust_vec_len, data_len),
-        70 => {
+        66 => wire__crate__api__network__net_party_active_impl(port, ptr, rust_vec_len, data_len),
+        67 => wire__crate__api__network__net_party_requests_impl(port, ptr, rust_vec_len, data_len),
+        68 => wire__crate__api__network__net_party_start_impl(port, ptr, rust_vec_len, data_len),
+        69 => wire__crate__api__network__net_party_stop_impl(port, ptr, rust_vec_len, data_len),
+        70 => wire__crate__api__network__net_party_update_impl(port, ptr, rust_vec_len, data_len),
+        71 => {
             wire__crate__api__network__net_pending_approvals_impl(port, ptr, rust_vec_len, data_len)
         }
-        71 => {
+        72 => {
             wire__crate__api__library__net_recent_transfers_impl(port, ptr, rust_vec_len, data_len)
         }
-        72 => wire__crate__api__library__net_remember_host_impl(port, ptr, rust_vec_len, data_len),
-        73 => wire__crate__api__network__net_revoke_all_impl(port, ptr, rust_vec_len, data_len),
-        74 => wire__crate__api__network__net_revoke_peer_impl(port, ptr, rust_vec_len, data_len),
-        75 => wire__crate__api__network__net_start_host_impl(port, ptr, rust_vec_len, data_len),
-        76 => wire__crate__api__network__net_stop_host_impl(port, ptr, rust_vec_len, data_len),
-        77 => {
+        73 => wire__crate__api__library__net_remember_host_impl(port, ptr, rust_vec_len, data_len),
+        74 => wire__crate__api__network__net_revoke_all_impl(port, ptr, rust_vec_len, data_len),
+        75 => wire__crate__api__network__net_revoke_peer_impl(port, ptr, rust_vec_len, data_len),
+        76 => wire__crate__api__network__net_start_host_impl(port, ptr, rust_vec_len, data_len),
+        77 => wire__crate__api__network__net_stop_host_impl(port, ptr, rust_vec_len, data_len),
+        78 => {
             wire__crate__api__library__playlist_add_tracks_impl(port, ptr, rust_vec_len, data_len)
         }
-        78 => wire__crate__api__library__playlist_create_impl(port, ptr, rust_vec_len, data_len),
-        79 => wire__crate__api__library__playlist_delete_impl(port, ptr, rust_vec_len, data_len),
-        80 => wire__crate__api__library__playlist_duplicate_impl(port, ptr, rust_vec_len, data_len),
-        81 => wire__crate__api__library__playlist_export_impl(port, ptr, rust_vec_len, data_len),
-        82 => wire__crate__api__library__playlist_import_impl(port, ptr, rust_vec_len, data_len),
-        83 => wire__crate__api__library__playlist_list_impl(port, ptr, rust_vec_len, data_len),
-        84 => wire__crate__api__library__playlist_remove_position_impl(
+        79 => wire__crate__api__library__playlist_create_impl(port, ptr, rust_vec_len, data_len),
+        80 => wire__crate__api__library__playlist_delete_impl(port, ptr, rust_vec_len, data_len),
+        81 => wire__crate__api__library__playlist_duplicate_impl(port, ptr, rust_vec_len, data_len),
+        82 => wire__crate__api__library__playlist_export_impl(port, ptr, rust_vec_len, data_len),
+        83 => wire__crate__api__library__playlist_import_impl(port, ptr, rust_vec_len, data_len),
+        84 => wire__crate__api__library__playlist_list_impl(port, ptr, rust_vec_len, data_len),
+        85 => wire__crate__api__library__playlist_remove_position_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        85 => wire__crate__api__library__playlist_rename_impl(port, ptr, rust_vec_len, data_len),
-        86 => wire__crate__api__library__playlist_reorder_tracks_impl(
+        86 => wire__crate__api__library__playlist_rename_impl(port, ptr, rust_vec_len, data_len),
+        87 => wire__crate__api__library__playlist_reorder_tracks_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        87 => wire__crate__api__library__playlist_tracks_impl(port, ptr, rust_vec_len, data_len),
-        88 => wire__crate__api__library__settings_delete_impl(port, ptr, rust_vec_len, data_len),
-        89 => wire__crate__api__library__settings_get_impl(port, ptr, rust_vec_len, data_len),
-        90 => wire__crate__api__library__settings_set_impl(port, ptr, rust_vec_len, data_len),
-        91 => wire__crate__api__library__share_list_impl(port, ptr, rust_vec_len, data_len),
-        92 => wire__crate__api__library__share_remove_impl(port, ptr, rust_vec_len, data_len),
-        93 => wire__crate__api__library__share_set_impl(port, ptr, rust_vec_len, data_len),
-        94 => wire__crate__api__library__share_set_enabled_impl(port, ptr, rust_vec_len, data_len),
-        95 => {
+        88 => wire__crate__api__library__playlist_tracks_impl(port, ptr, rust_vec_len, data_len),
+        89 => wire__crate__api__library__settings_delete_impl(port, ptr, rust_vec_len, data_len),
+        90 => wire__crate__api__library__settings_get_impl(port, ptr, rust_vec_len, data_len),
+        91 => wire__crate__api__library__settings_set_impl(port, ptr, rust_vec_len, data_len),
+        92 => wire__crate__api__library__share_list_impl(port, ptr, rust_vec_len, data_len),
+        93 => wire__crate__api__library__share_remove_impl(port, ptr, rust_vec_len, data_len),
+        94 => wire__crate__api__library__share_set_impl(port, ptr, rust_vec_len, data_len),
+        95 => wire__crate__api__library__share_set_enabled_impl(port, ptr, rust_vec_len, data_len),
+        96 => {
             wire__crate__api__library__smart_playlist_create_impl(port, ptr, rust_vec_len, data_len)
         }
-        96 => {
+        97 => {
             wire__crate__api__library__smart_playlist_delete_impl(port, ptr, rust_vec_len, data_len)
         }
-        97 => {
+        98 => {
             wire__crate__api__library__smart_playlist_list_impl(port, ptr, rust_vec_len, data_len)
         }
-        98 => wire__crate__api__library__smart_playlist_preview_impl(
+        99 => wire__crate__api__library__smart_playlist_preview_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        99 => {
+        100 => {
             wire__crate__api__library__smart_playlist_tracks_impl(port, ptr, rust_vec_len, data_len)
         }
-        100 => {
+        101 => {
             wire__crate__api__library__smart_playlist_update_impl(port, ptr, rust_vec_len, data_len)
         }
         _ => unreachable!(),